Tune hyperparameters
Tried different configs and logged them in Neptune and this one is
not bad. In general, architectures with batch normalization had
problems (pretty unstable on validation set). Alexnet was also good
(as much as this one or a custom vanilla CNN) but has way more params.
Also, this one was simpler to tune than a custom CNN so preferable