Track PTH with git LFS again but exclude danceability.pth
Still because of bandwidth quota. In general we want to track PTH
files because they may be very large but in this case we can do an
exception. Also, when the quota is reset we can still migrate it if
we want (e.g. if we plan to overwrite danceability.pth multiple times
or with a large file)