Hyperparameters tuning (#9)
* model_danceability.ipynb: adjust model reloading

* data_collection.ipynb: adjust module reloadings

* Tune hyperparameters

Tried different configs and logged them in Neptune and this one is
not bad. In general, architectures with batch normalization had
problems (pretty unstable on validation set). Alexnet was also good
(as much as this one or a custom vanilla CNN) but has way more params.
Also, this one was simpler to tune than a custom CNN so preferable

* Delete model params because they are of the old model

I still need to recreate the new ones (I ran out of the GPU hours)

* Temporarily untrack PTH to bypass Git LFS bandwidth quota

* Train a good danceability model

Hyperparameters used and training history are those currently in
model_danceability.ipynb (even if trained on Colab).
Model parameters (danceability.pth) are pushed as a regular file
because we reached the monthly limit of Git LFS bandwidth quota

* Track PTH with git LFS again but exclude danceability.pth

Still because of bandwidth quota. In general we want to track PTH
files because they may be very large but in this case we can do an
exception. Also, when the quota is reset we can still migrate it if
we want (e.g. if we plan to overwrite danceability.pth multiple times
or with a large file)

* Update README.md